update: 添加中间件错误处理
