+++ middleware/mi-log/index.js
@@ -1,77 +1,26 @@
-const log4js = require('log4js');
-const path = require("path");
-const client = require("./client.js");
-const logger = log4js.getLogger('xxxx');
-
-
-// ALL OFF 这两个等级并不会直接在业务代码中使用
-const methods = ["trace", "debug", "info", "warn", "error", "fatal", "mark"];
-// 几个常用的type
-// const types = ["console", "file", "dateFile", "levelFilter", "pattern"];
-
-const baseInfo = {
-  appLogLevel: 'debug',
-  filename: 'logs/task'
-}
+const convert = require("koa-convert");
+const logger = require("./logger");
 
 /**
- * log 中间件主函数
+ * 中间件错误处理
  * 
- * @param {string} env - 环境变量
- * @param {Object} options 
- * @param {Object} options.projectName - 项目名称
- * @param {Object} options.appLogLevel - app日志记录的日志级别
- * @param {Object} options.serverIp - 当前服务器ip
+ * @param {any} options 
  * @returns 
  */
-module.exports = (env, options) => {
-  const contextLogger = {};
-  const opts = Object.assign({}, baseInfo, options || {}) ;
-  const {
-    projectName,
-    serverIp,
-    filename,
-    appLogLevel
-  } = opts;
-  const currentLevel = methods.findIndex(ele => ele === appLogLevel)
-  const appenders = {};
-  const commonInfo = { projectName, serverIp };
+module.exports = (options) => {
 
-  appenders.tasks = {
-    type: 'dateFile',
-    filename,
-    pattern: '-yyyy-MM-dd.log',
-    alwaysIncludePattern: true
-  }
-  // 开发环境添加终端打印日志
-  if (env === "dev" || env === "local" || env === "development") {
-    appenders.out = {
-      type: "console"
-    }
-  }
+  const loggerMiddleware = convert(logger(options));
 
-  log4js.configure({
-    appenders,
-    categories: {
-      default: {
-        appenders: Object.keys(appenders),
-        level: appLogLevel
-      }
-    }
-  })
-  // 将log挂在上下文上
-  return async (ctx, next) => {
-    // level 以上级别的日志方法 
-    methods.forEach((method, i) => {
-      if (i >= currentLevel) {
-        contextLogger[method] = (message) => {
-          logger[method](client(ctx, message, commonInfo))
+  return (ctx, next) => {
+
+    return loggerMiddleware(ctx, next)
+    .catch((e) => {
+        if (ctx.status < 500) {
+            ctx.status = 500;
         }
-      } else {
-        contextLogger[method] = () => {}
-      }
+        ctx.log.error(e.stack);
+        ctx.state.logged = true;
+        ctx.throw(e);
     });
-    ctx.log = contextLogger;
-    await next()
   };
 }
